pyproject.toml
[tool.poetry]
name = "tilemap"
version = "0.1.0"
description = ""
authors = ["Andrey Raspopov <raspopov@arrival.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
